+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+- OpenRouter: Retry 500 and 504 errors returned in request JSON body.
+- Scoring: Allow customisation of grouped metric names.
+- Inspect View: Show live requests and responses for model call events in the transcript.
+- Inspect View: Improve scroll performance when viewing sample transcripts and messages.
+
 ## 0.3.172 (06 February 2026)
 
 - Inspect View: Fix a regression which affected the display of samples within VSCode.